+++ FOX/__init__.py
@@ -14,18 +14,10 @@ https://auto-fox.readthedocs.io/en/latest/
 from .__version__ import __version__
 
 from .functions import (
-    get_rdf_lowmem, get_rdf,
-    get_adf,
-    get_template, assert_error, get_example_xyz,
-    update_charge,
-    estimate_lj,
-    get_non_bonded
+    assert_error, get_example_xyz
 )
 
 from .io import (
-    read_kf,
-    read_multi_xyz,
-    read_pdb,
     PSFContainer,
     PRMContainer,
     create_hdf5, create_xyz_hdf5, to_hdf5, from_hdf5
@@ -37,20 +29,21 @@ from .classes import (
     ARMC, run_armc,
 )
 
-__version__ = __version__
+from .ff import (
+    get_bonded,
+    get_non_bonded,
+    get_intra_non_bonded,
+    estimate_lj
+)
+
 __author__ = "Bas van Beek"
 __email__ = 'b.f.van.beek@vu.nl'
 
 __all__ = [
-    'get_rdf_lowmem', 'get_rdf',
-    'get_adf',
-    'get_template', 'assert_error', 'get_example_xyz',
-    'estimate_lj',
-    'get_non_bonded',
+    '__version__',
 
-    'read_kf',
-    'read_multi_xyz',
-    'read_pdb',
+    'get_example_xyz', 'assert_error',
+
     'PSFContainer',
     'PRMContainer',
     'create_hdf5', 'create_xyz_hdf5', 'to_hdf5', 'from_hdf5',
@@ -58,4 +51,9 @@ __all__ = [
     'FrozenSettings',
     'MultiMolecule',
     'ARMC', 'run_armc',
+
+    'estimate_lj',
+    'get_non_bonded',
+    'get_intra_non_bonded',
+    'get_bonded'
 ]