+++ FOX/ff/lj_calculate.py
@@ -1,8 +1,22 @@
-"""
-FOX.functions.lj_calculate
-==========================
+r"""
+FOX.ff.lj_calculate
+===================
 
 A module for calculating non-bonded interactions using Coulomb + Lennard-Jones potentials.
+
+.. math::
+
+    V_{LJ} = 4 \varepsilon
+    \left(
+        \left(
+            \frac{\sigma}{r}
+        \right )^{12} -
+        \left(
+            \frac{\sigma}{r}
+        \right )^6
+    \right )
+
+    V_{Coulomb} = \frac{1}{4 \pi \varepsilon_{0}} \frac{q_{i} q_{j}}{r_{ij}}
 
 """
 
@@ -17,7 +31,7 @@ import pandas as pd
 
 from scm.plams import Units, Settings
 
-from .utils import read_rtf_file, fill_diagonal_blocks, group_by_values
+from ..functions.utils import read_rtf_file, fill_diagonal_blocks, group_by_values
 from ..classes.multi_mol import MultiMolecule
 from ..io.read_psf import PSFContainer
 from ..io.read_prm import PRMContainer