+++ FOX/ff/lj_param.py
@@ -1,6 +1,6 @@
 """
-FOX.functions.lj_param
-======================
+FOX.ff.lj_param
+===============
 
 A module for estimating Lennard-Jones parameters.
 
@@ -169,7 +169,7 @@ def estimate_lj(rdf: pd.DataFrame, temperature: float = 298.15,
     lj_dict: Dict[str, List[float]] = {'sigma (Angstrom)': []}
     sigma_append = lj_dict['sigma (Angstrom)'].append
 
-    for atoms, distr in rdf.items():
+    for distr in rdf.values:
         if sigma_estimate == 'inflection':
             grad = np.gradient(distr)
             i = np.argmax(grad)  # The first inflection point in the RDF