+++ FOX/functions/molecule_utils.py
@@ -23,7 +23,7 @@ from typing import List
 
 import numpy as np
 
-from scm.plams import Molecule
+from scm.plams import Molecule, Atom
 
 __all__ = ['get_bonds', 'get_angles', 'get_dihedrals', 'get_impropers']
 
@@ -72,13 +72,13 @@ def separate_mod(mol: Molecule) -> List[List[int]]:
         at._visited = False
 
     # Loop through atoms
-    def dfs(at1, m):
+    def dfs(at1: Atom, m_append: list.append):
         at1._visited = True
-        m.append(at1.id)
+        m_append(at1.id)
         for bond in at1.bonds:
             at2 = bond.other_end(at1)
             if not at2._visited:
-                dfs(at2, m)
+                dfs(at2, m_append)
 
     # Create a nested list of atomic indices
     indices = []
@@ -86,7 +86,7 @@ def separate_mod(mol: Molecule) -> List[List[int]]:
     for at in mol.atoms:
         if not at._visited:
             m = []
-            dfs(at, m)
+            dfs(at, m.append)
             indices_append(m)
 
     return indices
@@ -139,6 +139,7 @@ def get_angles(mol: Molecule) -> np.ndarray:
     """
     mol.set_atoms_id()
     angle = []
+    angle_append = angle.append
 
     for at2 in mol.atoms:
         if len(at2.bonds) < 2:
@@ -147,7 +148,7 @@ def get_angles(mol: Molecule) -> np.ndarray:
         at_other = [bond.other_end(at2) for bond in at2.bonds]
         for i, at1 in enumerate(at_other, 1):
             for at3 in at_other[i:]:
-                angle.append((at1.id, at2.id, at3.id))
+                angle_append((at1.id, at2.id, at3.id))
 
     ret = np.array(angle, dtype=int, ndmin=2)
     mol.unset_atoms_id()
@@ -155,8 +156,8 @@ def get_angles(mol: Molecule) -> np.ndarray:
         return ret
 
     # Sort horizontally
-    idx1 = np.argsort(ret[:, [0, -1]], axis=1)
-    ret[:, 0], ret[:, -1] = np.take_along_axis(ret[:, [0, -1]], idx1, axis=1).T
+    idx1 = np.argsort(ret[:, ::2], axis=1)
+    ret[:, ::2] = np.take_along_axis(ret[:, ::2], idx1, axis=1)
 
     # Sort and return vertically
     idx2 = np.argsort(ret, axis=0)[:, 0]
@@ -179,6 +180,7 @@ def get_dihedrals(mol: Molecule) -> np.ndarray:
     """
     mol.set_atoms_id()
     dihed = []
+    dihed_append = dihed.append
 
     for b1 in mol.bonds:
         if not (len(b1.atom1.bonds) > 1 and len(b1.atom2.bonds) > 1):
@@ -193,16 +195,12 @@ def get_dihedrals(mol: Molecule) -> np.ndarray:
             for b3 in at3.bonds:
                 at4 = b3.other_end(at3)
                 if at4 != at2:
-                    dihed.append((at1.id, at2.id, at3.id, at4.id))
+                    dihed_append((at1.id, at2.id, at3.id, at4.id))
 
     ret = np.array(dihed, dtype=int, ndmin=2)
     mol.unset_atoms_id()
     if not dihed:  # If no dihedrals are found
         return ret
-
-    # Sort horizontally
-    idx1 = np.argsort(ret[:, [0, -1]], axis=1)
-    ret[:, 0], ret[:, -1] = np.take_along_axis(ret[:, [0, -1]], idx1, axis=1).T
 
     # Sort and return vertically
     idx2 = np.argsort(ret, axis=0)[:, 0]
@@ -225,6 +223,7 @@ def get_impropers(mol: Molecule) -> np.ndarray:
     """
     mol.set_atoms_id()
     impropers = []
+    impropers_append = impropers.append
 
     for at1 in mol.atoms:
         order = [bond.order for bond in at1.bonds]
@@ -233,7 +232,7 @@ def get_impropers(mol: Molecule) -> np.ndarray:
 
         if 2.0 in order or 1.5 in order:
             at2, at3, at4 = [bond.other_end(at1) for bond in at1.bonds]
-            impropers.append((at1.id, at2.id, at3.id, at4.id))
+            impropers_append((at1.id, at2.id, at3.id, at4.id))
 
     ret = np.array(impropers, dtype=int, ndmin=2)
     mol.unset_atoms_id()
@@ -243,5 +242,5 @@ def get_impropers(mol: Molecule) -> np.ndarray:
     # Sort along the rows of columns 2, 3 & 4 based on atomic mass in descending order
     mass = np.array([[mol[int(j)].mass for j in i] for i in ret[:, 1:]])
     idx = np.argsort(mass, axis=1)[:, ::-1]
-    ret[:, 1], ret[:, 2], ret[:, 3] = np.take_along_axis(ret[:, 1:], idx, axis=1).T
+    ret[:, 1:] = np.take_along_axis(ret[:, 1:], idx, axis=1)
     return ret